fix(dual-screen): resolve bottom-screen display bugs and game launch freeze
- Corrected native Java WebView bridge pushFrame signature to accept double types, eliminating silent reflection matching failures and preventing stale/frozen presentation views.
- Modified active/ghost piece mappings in the game loop to compute relative grid row offsets via py - 20, keeping the falling pieces and ghost blocks visible on the lower active field view even after screen flips.
- Added a try...catch safety wrapper in the JS dual-screen HUD push loop to prevent exceptions from crashing the main game loop thread.
- Fixed a ReferenceError crash in getMenuStatePayload where undefined mode-specific gravity variables were accessed, changing them to use the global startingGravity variable.

+++ android/app/src/main/java/com/timetoady/tetrisflip/DualScreenHudPlugin.java
@@ -65,15 +65,15 @@ public class DualScreenHudPlugin extends Plugin {
                 if (getBridge() != null && getBridge().getWebView() != null) {
                     getBridge().getWebView().addJavascriptInterface(new Object() {
                         @JavascriptInterface
-                        public void pushFrame(String cells, int score, int level, int lines, String status, boolean isFlipped) {
-                            pushFrame(cells, score, level, lines, status, isFlipped, 0f, 0f, 0f);
+                        public void pushFrame(String cells, double score, double level, double lines, String status, boolean isFlipped) {
+                            pushFrame(cells, score, level, lines, status, isFlipped, 0.0, 0.0, 0.0);
                         }
 
                         @JavascriptInterface
-                        public void pushFrame(String cells, int score, int level, int lines, String status, boolean isFlipped, float cellSize, float gridLeft, float gridTop) {
+                        public void pushFrame(String cells, double score, double level, double lines, String status, boolean isFlipped, double cellSize, double gridLeft, double gridTop) {
                             getActivity().runOnUiThread(() -> {
                                 if (controller != null && controller.hasPresentation()) {
-                                    controller.pushFrame(cells, score, level, lines, status, isFlipped, cellSize, gridLeft, gridTop);
+                                    controller.pushFrame(cells, (int) score, (int) level, (int) lines, status, isFlipped, (float) cellSize, (float) gridLeft, (float) gridTop);
                                 }
                             });
                         }
